Move and rename DataConsistencyChecker
